refactor: Provider에서 DataAction, reducer 분리
