--- src/contexts/context.constant.ts
@@ -1,6 +1,0 @@
-import { ErrorState } from '../components/ErrorToast/error.type';
-
-export const INITIAL_ERROR: ErrorState = {
-  is: false,
-  message: '',
-};
